# backend/api/openapi.yaml
openapi: "3.1.0"
info: {title: Oh-My-AIHub API, version: "0.1.0", description: "Oh-My-AIHub 前后端契约的唯一来源（见 ADR）。覆盖全部 /api/** JSON 接口；/v1/**、/v1beta/** 外部模型协议入口只登记路径与认证。\n\n约定：金额一律为十进制字符串；时间为 RFC 3339；错误统一为 {\"error\":{\"code\",\"message\"}}；除 GET/HEAD/OPTIONS 外的 /api 请求须携带与 Host 同源的 Origin。\n\n扩展字段 x-access 声明路由门禁，由契约测试与路由表逐项对照：public 无需会话；session 需会话（允许首次改密前访问）；ready 需会话且已完成首次改密；admin 在 ready 基础上需管理员；gateway_key 凭网关 API Key。"}
servers:
  - url: /
    description: "同源部署：开发环境由 Vite 代理，Compose 环境由 Nginx 代理。"
...
paths:
...
  /api/channels:
    get:
      operationId: listChannels
      tags: [channels]
      summary: "我的渠道"
      x-access: ready
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelList"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: createChannel
      tags: [channels]
      summary: "创建渠道"
      x-access: ready
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateChannelRequest"}
      responses:
        '201':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channels/{channelID}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: "我的渠道详情"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: "更新渠道"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateChannelRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: "删除渠道"
      description: "DELETE 请求携带 JSON 请求体（expected_version）。"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channels/{channelID}/publish:
    post:
      operationId: publishChannel
      tags: [channels]
      summary: "发布渠道"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channels/{channelID}/pause:
    post:
      operationId: pauseChannel
      tags: [channels]
      summary: "暂停渠道"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channels/{channelID}/credential-revoke:
    post:
      operationId: revokeChannelCredential
      tags: [channels]
      summary: "撤销渠道凭据"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerChannelEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channels/{channelID}/offers:
    post:
      operationId: addChannelOffer
      tags: [channels]
      summary: "新增报价"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/channelID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AddOfferRequest"}
      responses:
        '201':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerOfferEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channel-offers/{offerID}:
    patch:
      operationId: updateChannelOffer
      tags: [channels]
      summary: "更新报价"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/offerID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateOfferRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerOfferEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
    delete:
      operationId: deleteChannelOffer
      tags: [channels]
      summary: "删除报价"
      description: "DELETE 请求携带 JSON 请求体（expected_version）。"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/offerID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerOfferEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channel-offers/{offerID}/disable:
    post:
      operationId: disableChannelOffer
      tags: [channels]
      summary: "停用报价"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/offerID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerOfferEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
  /api/channel-offers/{offerID}/resume:
    post:
      operationId: resumeChannelOffer
      tags: [channels]
      summary: "恢复报价"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/offerID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VersionedRequest"}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/OwnerOfferEnvelope"}
        '400': {$ref: "#/components/responses/BadRequest"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '404': {$ref: "#/components/responses/NotFound"}
        '409': {$ref: "#/components/responses/Conflict"}
        '422': {$ref: "#/components/responses/Unprocessable"}
...
  /api/market/offers:
    get:
      operationId: listMarketOffers
      tags: [market]
      summary: "市场报价列表"
      x-access: ready
      security:
        - sessionCookie: []
      parameters:
        - name: model_id
          in: query
          required: false
          description: "模型 ID。"
          schema: {type: string}
        - name: protocol
          in: query
          required: false
          description: "协议。"
          schema: {type: string, enum: [openai_chat_completions, openai_responses, anthropic_messages, google_gemini_generate_content]}
        - name: owner
          in: query
          required: false
          description: "按所有者名称搜索。"
          schema: {type: string}
        - name: sort
          in: query
          required: false
          description: "排序方式，缺省为 input_price。"
          schema:
            type: string
            enum: [input_price, output_price, cache_write_price, cache_read_price, success_rate, ttft, tps]
        - name: after
          in: query
          required: false
          description: "分页游标。"
          schema: {type: string}
        - name: limit
          in: query
          required: false
          description: "返回条数。"
          schema: {type: integer, default: 20}
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MarketOfferPage"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
        '422': {$ref: "#/components/responses/Unprocessable"}
...
  /api/admin/channels:
    get:
      operationId: listAdminChannels
      tags: [admin-channels]
      summary: "管理员：渠道列表"
      x-access: admin
      security:
        - sessionCookie: []
      responses:
        '200':
          description: "成功"
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminChannelList"}
        '401': {$ref: "#/components/responses/Unauthorized"}
        '403': {$ref: "#/components/responses/Forbidden"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oma_session
      description: "登录后由服务端设置的 HttpOnly 会话 Cookie。"
    bearerKey: {type: http, scheme: bearer, description: "平台 API Key：Authorization: Bearer <key>（OpenAI 协议）。"}
    anthropicKey:
      type: apiKey
      in: header
      name: x-api-key
      description: "平台 API Key（Anthropic 协议）。"
    geminiKey:
      type: apiKey
      in: header
      name: x-goog-api-key
      description: "平台 API Key（Gemini 协议）。"
...
  responses:
...
    Unauthorized:
      description: "未登录或会话失效（authentication_required）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
    Forbidden:
      description: "被拒绝：跨站请求（cross_origin_request）、首次登录须改密（password_change_required）、需要管理员（administrator_required）或无权操作（forbidden）。"
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...
  schemas:
...
    OwnerChannelList:
      type: object
      properties:
        channels:
          type: array
          items: {$ref: "#/components/schemas/OwnerChannel"}
      required: [channels]
      additionalProperties: false
...